Fix cancelling of builds.
This commit is proposed as a solution to https://github.com/leptos-rs/cargo-leptos/issues/54

The `build_proj` function did not consider the Outcome of each compile stage. This function has been altered so the build is short circuited if a compile is stopped.

Likewise the `build_all` function would build a list of projects without checking for shutdowns. This function has been changed to check after a call to `build_proj` to also short circuit if a shutdown is requested.

build_proj is called from end2end and at the start of a watch.  Both these modules have been changed so that after `build_proj` has been called they check for shutdown requests and exit if required.